+++ features/components/resources/download.steps.js
@@ -1,4 +1,4 @@
-//features/components/resources/downloadSteps.js
+//features/components/resources/download.steps.js
 var page = require('./download.page.js');
 var {defineSupportCode} = require('cucumber');
 
@@ -8,18 +8,32 @@ defineSupportCode(({Given, When, Then}) => {
         return page.go();
     });
 
+    When('I change the download file select box to {stringInDoubleQuotes}', function (file) {
+        return page.selectDownloadFile(file);
+    });
+
     Then('I should see \'Download the CHPL\' as the page title', function () {
-        var pageTitle = page.downloadPage.pageTitle;
+        var pageTitle = page.model.pageTitle;
         return (expect(pageTitle.getText()).eventually.to.equal('Download the CHPL'));
     });
 
-    Then('There should be {int} Download Files', function (count) {
-        var downloadOptions = page.downloadPage.downloadOptions;
+    Then('there should be {int} Download Files', function (count) {
+        var downloadOptions = page.model.downloadOptions;
         return (expect(downloadOptions.count()).to.eventually.equal(count));
     });
 
-    Then('There should be {int} Definition Files', function (count) {
-        var definitionOptions = page.downloadPage.definitionOptions;
+    Then('there should be {int} Definition Files', function (count) {
+        var definitionOptions = page.model.definitionOptions;
         return (expect(definitionOptions.count()).to.eventually.equal(count));
+    });
+
+    Then('the download select box should be {stringInDoubleQuotes}', function (file) {
+        var activeDownloadFile = page.model.activeDownloadFile;
+        return (expect(activeDownloadFile.getText()).to.eventually.equal(file));
+    });
+
+    Then('the definition select box should be {stringInDoubleQuotes}', function (file) {
+        var activeDefinitionFile = page.model.activeDefinitionFile;
+        return (expect(activeDefinitionFile.getText()).to.eventually.equal(file));
     });
 });