+++ protractor.conf.js
@@ -30,9 +30,9 @@ exports.config = {
         ],
         // Only execute the features or scenarios with tags matching @dev.
         // This may be an array of strings to specify multiple tags to include.
-        tags: ['~@ignore','~@wip'],
+        tags: ['~@ignore'],
         // How to format features (progress, summary, pretty, json)
-        format: ['progress', 'pretty', 'json:test_reports/cucumber_report.json'],
+        format: ['progress', /*'pretty', */'json:test_reports/cucumber_report.json'],
     },
 
     // Options to be passed to Jasmine-node.